+++ verify/error.go
@@ -1,10 +1,19 @@
 package verify
 
-// TODO: type embed comparable (err) and string (err.String())
+// FluentError encapsulates assertions for error object.
+type FluentError struct {
+	FluentObj[error]
+	FluentString[string]
+}
 
-// TODO: NotNil
-
-// TODO: Nil
+// Error is used for testing error object.
+func Error(got error) FluentError {
+	res := FluentError{FluentObj: FluentObj[error]{got}}
+	if got != nil {
+		res.FluentString.Got = got.Error()
+	}
+	return res
+}
 
 // TODO: Is
 
@@ -14,6 +23,6 @@ package verify
 
 // TODO: NotAs
 
-// TODO: Unwrappable
+// TODO: WithWrapped
 
-// TODO: NotUnwrappable
+// TODO: NoWrapped